+++ tests/testthat/test_helpers.R
@@ -9,3 +9,16 @@ test_that(".checkAlts functions correctly", {
     gr$ALT <- "A"
     expect_equal(gr, .checkAlts(gr, "ALT"))
 })
+
+test_that(".pasreVariants functions correctly",{
+    library(VariantAnnotation)
+    f <- VcfFile(
+        system.file("extdata/1000GP_subset.vcf.gz", package = "transmogR")
+    )
+    gr <- .parseVariants(f, "ALT", GRanges("chr1:1-500000"))
+    expect_equal(length(gr), 14L)
+    expect_true(all(colnames(mcols(gr)) == c("REF", "ALT")))
+    expect_true(is(gr$REF, "character"))
+    expect_true(is(gr$ALT, "character"))
+
+})